split database into seperate collection commands and added streams to proto
